* dev: (47 commits)
  IDE 2.5.2 VM 411
  Translated using Weblate (Czech)
  Translated using Weblate (Spanish)
  Translated using Weblate (German)
  Translated using Weblate (Catalan)
  Tweak to make Boardie compiler happy
  Try experimental NeoPixel implementation for ESP32 boards (except ESP32-C3) in attempt to reduce glitchs in ultra picky NeoPixel strips
  Update README file
  Boardie DAC primitive now works for buffers of 1024 or larger. Needs rework deal with single-byte writes and flow control.
  Fix build on boards without a TFT, and a missing return in primRotateDisplay
  WebPanel: remove a duplicated spec line
  BLE Keyboard: fix unbalanced quote in the key modifier menu
  Changes to electron package (Google says these should definitely NOT be .gitignored)
  Revert Bitbucket README.md change
  README.md edited online with Bitbucket
  README.md edited online with Bitbucket
  README.md edited online with Bitbucket
  rotate display primitive -useful if you want to enclose the board in a case in a different orientation-
  fix: DFU uploader working in Electron now
  added Luis Mateos's Telegram library. Thank you!
  ...

+++ vm/ioPrims.cpp
@@ -3495,7 +3495,7 @@ static OBJ primSquareWave(int argCount, OBJ *args) {
 	}
 
 	extern "C" void deepSleep(int secs) {
-		uint64_t usecs = 1000000 * secs;
+		uint64_t usecs = 1000000 * (uint64_t) secs;
 		uint64_t maxSleep = ESP.deepSleepMax() - 10000;
 		if (usecs > maxSleep) usecs = maxSleep;
 		ESP.deepSleep(usecs);
